fix review point C5

+++ core/rust/qdbr/parquet2/src/write/file.rs
@@ -653,7 +653,7 @@ impl<W: Write> ParquetFile<W> {
     pub fn replace<E>(
         &mut self,
         row_group: RowGroupIter<'_, E>,
-        ordinal: Option<i16>,
+        ordinal: Option<i32>,
         bloom_hashes: &[Option<Arc<Mutex<HashSet<u64>>>>],
     ) -> std::result::Result<(), E>
     where
@@ -689,7 +689,7 @@ impl<W: Write> ParquetFile<W> {
     pub fn insert<E>(
         &mut self,
         row_group: RowGroupIter<'_, E>,
-        position: i16,
+        position: i32,
         bloom_hashes: &[Option<Arc<Mutex<HashSet<u64>>>>],
     ) -> std::result::Result<(), E>
     where
